Refactor tests and add case for non-generator-class with convention name
