--- core/src/test/java/com/pholser/junit/quickcheck/internal/generator/BrokenConvention.java
@@ -1,5 +1,0 @@
-package com.pholser.junit.quickcheck.internal.generator;
-
-public class BrokenConvention {
-    // just need an object in a package
-}
