+++ core/src/test/java/com/pholser/junit/quickcheck/internal/generator/conventiontestclasses/Convention.java
@@ -1,0 +1,5 @@
+package com.pholser.junit.quickcheck.internal.generator.conventiontestclasses;
+
+public class Convention {
+    // just need an object in a package
+}
