+++ core/src/test/java/com/pholser/junit/quickcheck/internal/generator/conventiontestclasses/GeneratesOtherTypes.java
@@ -1,0 +1,5 @@
+package com.pholser.junit.quickcheck.internal.generator.conventiontestclasses;
+
+public class GeneratesOtherTypes {
+    // just need an object in a package
+}
