+++ core/src/test/java/com/pholser/junit/quickcheck/internal/generator/conventiontestclasses/NotAGenerator.java
@@ -1,0 +1,5 @@
+package com.pholser.junit.quickcheck.internal.generator.conventiontestclasses;
+
+public class NotAGenerator {
+    // just need an object in a package
+}
