+++ core/src/test/java/com/pholser/junit/quickcheck/internal/generator/conventiontestclasses/NotAGeneratorGen.java
@@ -1,0 +1,5 @@
+package com.pholser.junit.quickcheck.internal.generator.conventiontestclasses;
+
+public class NotAGeneratorGen {
+    // just need an object in a package with this name
+}
